Change int result to long and refactor

+++ src/main/java/com/ail/crxmarkets/algorithm/RainyHillsAlgorithm.java
@@ -2,6 +2,6 @@ package com.ail.crxmarkets.algorithm;
 
 public interface RainyHillsAlgorithm {
 
-	int calcWaterVolumeOnSurface(int[] surface);
+	long calcWaterVolumeOnSurface(int[] surface);
 
 }
